examples/graphiti/requirements.txt
graphiti-core>=0.3.0
pyyaml>=6.0
